requirements.txt
# Main Framework
llama-index-core==0.14.12
llama-index==0.14.12

# OpenAI Plugins (Updated for Core 0.14)
llama-index-llms-openai==0.6.13
llama-index-embeddings-openai==0.5.1
openai==1.109.1

# Vector Store & Parsing
llama-index-vector-stores-chroma==0.5.5
chromadb==1.4.1
llama-parse==0.6.54

# Utilities
python-dotenv==1.0.1
streamlit==1.53.0